fix #22
