+++ web/.gitignore
@@ -1,3 +1,4 @@
 node_modules
 .cache
 dist
+coverage
