+++ web/jest.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  collectCoverageFrom: ['src/**/*.{ts,tsx}'],
+  globals: {
+    'ts-jest': {},
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
+  testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$',
+  transform: {
+    '^.+\\.tsx?$': 'ts-jest',
+  },
+  transformIgnorePatterns: ['/node_modules/'],
+}
